move tree keys to verkle and simplify
