+++ packages/verkle/src/index.ts
@@ -2,4 +2,5 @@ export * from './db/index.js'
 export * from './node/index.js'
 export * from './types.js'
 export * from './util/index.js'
+export * from './util/keys.js'
 export * from './verkleTree.js'
